% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5.temp.rh.wind.extract.R
\name{era5.temp.rh.wind.extract}
\alias{era5.temp.rh.wind.extract}
\title{Temperature, RH, and Wind extraction}
\usage{
era5.temp.rh.wind.extract(latitude, longitude, date.start, date.end)
}
\arguments{
\item{latitude}{Latitude of point (it will find closest ERA5 grid point)}

\item{longitude}{Longitude of point}

\item{date.start}{Start of ERA5 data---NOT date of interest---date in file}

\item{date.end}{As above}
}
\description{
This function allows you to extract out just the Temperature, RH, and Wind from ERA5 data---can be used with map to do many points on a grid (but beware of too many files open)
}
\examples{
dry.2015 <- map2_dfr(grid.subset2$Var1, grid.subset2$Var2, era5.temp.rh.wind.extract, date.start = "2015-03-01", date.end="2015-05-31")
This is pulling out all temps, RH, and winds (converting to u/v wind) for each point in the grid from every night of interest.
}
\keyword{ERA5}
