% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/era5extractsingle.R
\name{era5extractsingle}
\alias{era5extractsingle}
\title{Single level ERA5 extraction.}
\usage{
era5extractsingle(file, variable, lat, long)
}
\arguments{
\item{file}{Filename}

\item{variable}{ERA5 variable name, in my case t2m (temp 2 meters) or tp (total precip)}

\item{lat}{Latitude of point (it will find closest ERA5 grid point)}

\item{long}{Longitude of point}
}
\description{
This is a (mostly helper) function for extracting out single level (height) era5 data
}
\examples{
era5extractsingle("t2m_201803.nc", "t2m", 12.5, 11.5)
}
\keyword{ERA5}
