% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.to.polygon.R
\name{grid.to.polygon}
\alias{grid.to.polygon}
\title{Subset ERA5 grid to polygon}
\usage{
grid.to.polygon(ncfile, polygon.df)
}
\arguments{
\item{ncfile}{ncfile to get grid --- use a multilevel file or change the lat/long names}

\item{polygon.df}{Dataframe of polygon with columns lat and long}
}
\description{
This is a function to find the extents of a polygon, find the closest grid points (i.e. from ERA5) inside these extents and output the data
}
\examples{
koulikoro.grid <- subset.grid.to.polygon(ncfile="2018/wd_201803.nc", polygon.df = out_df )
}
\keyword{ERA5}
